requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pydot==4.0.1
PyJWT==2.9.0
pyparsing==3.2.3
sqlparse==0.5.3
typing_extensions==4.13.2